refactor: notice 텍스트편집기 컴포넌트 분리 및 폴더링
