+++ src/styles/index.css
@@ -1,1 +1,76 @@
 @import url('https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard-dynamic-subset.css');
+
+/* TipTap 목록 스타일 - 수정된 버전 */
+.ProseMirror ul {
+  list-style-type: disc;
+  list-style-position: outside;
+  margin-left: 1.5rem;
+  padding-left: 0;
+  margin-top: 0.5rem;
+  margin-bottom: 0.5rem;
+}
+
+.ProseMirror ol {
+  list-style-type: decimal;
+  list-style-position: outside;
+  margin-left: 1.5rem;
+  padding-left: 0;
+  margin-top: 0.5rem;
+  margin-bottom: 0.5rem;
+}
+
+.ProseMirror li {
+  display: list-item;
+  margin: 0.25rem 0;
+  padding-left: 0.5rem;
+  position: relative;
+  min-height: 1.2em;
+}
+
+/* 중첩된 목록을 위한 스타일 */
+.ProseMirror ul ul,
+.ProseMirror ol ol,
+.ProseMirror ul ol,
+.ProseMirror ol ul {
+  margin-top: 0.25rem;
+  margin-bottom: 0.25rem;
+}
+
+/* 커스텀 클래스 스타일 - 수정된 버전 */
+.tiptap-bullet-list {
+  list-style-type: disc !important;
+  list-style-position: outside !important;
+  margin-left: 1.5rem !important;
+  padding-left: 0 !important;
+}
+
+.tiptap-ordered-list {
+  list-style-type: decimal !important;
+  list-style-position: outside !important;
+  margin-left: 1.5rem !important;
+  padding-left: 0 !important;
+}
+
+.tiptap-list-item {
+  display: list-item !important;
+  margin: 0.25rem 0 !important;
+  padding-left: 0.5rem !important;
+  position: relative !important;
+  min-height: 1.2em !important;
+}
+
+/* 목록 항목 내부의 p 태그 스타일 조정 */
+.ProseMirror li p {
+  margin: 0;
+  padding: 0;
+}
+
+/* 포커스 및 커서 스타일 */
+.ProseMirror li:focus {
+  outline: none;
+}
+
+.ProseMirror li::marker {
+  color: #333;
+}
+
